flag para flows

+++ Mobilidade.py
@@ -410,30 +410,20 @@ def start_teste(args):
 
         put('http://127.0.0.1:8008/topology/json', data=dumps(topo))
     
-    log_dir = "./test"
-    
-    # creating log directory
-    # log_dir = path.expanduser('~') + log_dir
-    # i = 1
-    # while True:
-    #    if not path.exists(log_dir + str(i)):
-    #        # mkdir(log_dir + str(i))
-    #        log_dir = log_dir + str(i)
-    #        break
-    #    i = i+1
-    
-    experiment_duration = int(input("Experiment duration: "))
-    n_elephant_flows = int(input("No of elephant flows: "))
-    n_mice_flows = int(input("No of mice flows: "))
+    if '-f' in args:
+      log_dir = "./test"    
+      experiment_duration = int(input("Experiment duration: "))
+      n_elephant_flows = int(input("No of elephant flows: "))
+      n_mice_flows = int(input("No of mice flows: "))
 
-    # Inicia o log tcp dump
-    coleta_tcpdum(net)
+      # Inicia o log tcp dump
+      coleta_tcpdum(net)
     
-    # Gera e executa os flows
-    generate_flows(n_elephant_flows, n_mice_flows, experiment_duration, net, log_dir)
+      # Gera e executa os flows
+      generate_flows(n_elephant_flows, n_mice_flows, experiment_duration, net, log_dir)
     
-    # Encerra o log tcpdump
-    net.get('sw5').cmdPrint('killall tcpdump')
+      # Encerra o log tcpdump
+      net.get('sw5').cmdPrint('killall tcpdump')
 
     info("*** Rodando CLI \n")
     CLI(net)
